Enable --output for non Win32 platforms
Though it was documented for all builds of iperf, the -o, --output command line
option to redirect the output to a file was only complied in for Win32. This
patch enables it for Unix builds as well. All stdout, including the header, can
now be redirected to the specified file, leaving iperf completely silent on the
console apart from stderr.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>